+++ examples/react/fwoosh.config.ts
@@ -6,11 +6,11 @@ export const config: FwooshOptions = {
   title: "@fwoosh/react",
   setup: path.resolve("./config/fwoosh-setup.ts"),
   plugins: [
-    "@fwoosh/props-panel",
-    "@fwoosh/source-panel",
-    "@fwoosh/actions",
-    "@fwoosh/zoom",
-    "@fwoosh/measure",
+    "@fwoosh/panel-props",
+    "@fwoosh/panel-source",
+    "@fwoosh/panel-actions",
+    "@fwoosh/tool-zoom",
+    "@fwoosh/tool-measure",
     new ReactPlugin({
       docgenOptions: {
         propFilter: (prop) => {